refactor: Canvas, SubSection
