+++ src/components/pages/Generator/CanvasSection/Canvas/index.ts
@@ -1,0 +1,1 @@
+export {Canvas} from './Canvas';
