+++ src/components/pages/Generator/CanvasSection/SubSection/index.ts
@@ -1,0 +1,1 @@
+export {SubSection} from './SubSection';
